feat(fr_meteofrance): add the climatological-archive refresh flow
quotidienne, mensuelle and poste were onboarded but no flow refreshed them.
Adds fr_meteofrance_climatologie_base_flow, monthly at 51 7 8,9,10,11 (a minute
nothing else in the repo uses).

It downloads every period slice but re-cleans only latest-<years>. Météo-France
rewrites that slice alone, and staging objects are named <dept>_<period>.parquet,
so uploading it overwrites those objects in place and leaves the two historical
slices untouched -- 4M daily rows a month instead of 137M. The full download is
still needed because poste is the union over both series and all slices; built
from the refreshed slice alone it would drop stations that stopped reporting.

Per the DRY rule the transform now has one copy, in
pipelines/datasets/fr_meteofrance/{clim_schema,clim_utils}.py; the models/ CLIs
import it. Verified byte-identical on departement 01: all six parquet files
match the validated output exactly, and the vectorised geolocalisation matches
the uploaded register on all 14,751 rows.

Moving the code into pipelines/ also put it under pyrefly for the first time
(models/<ds>/code is excluded), surfacing two real type errors -- a df.apply
returning str|None, and pc.max missing from the stubs. Both fixed; the max
observation date now comes from parquet row-group statistics, reading no rows.
