+++ models/fr_meteofrance/code/clim_gen_artifacts.py
@@ -14,7 +14,7 @@ import sys
 from pathlib import Path
 
 sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-import clim_schema as cs
+from pipelines.datasets.fr_meteofrance import clim_schema as cs
 
 HERE = Path(os.path.dirname(os.path.abspath(__file__)))
 MODELS = HERE.parent
