+++ pipelines/datasets/fr_meteofrance/clim_schema.py
